session01/organized_approach/requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==1.0.2
opencv-python==4.5.3.56
pyyaml==5.4.1
joblib==1.0.1

# Development dependencies
pytest==6.2.4
black==21.6.0
flake8==3.9.2